renamed SavePhotoDB to PhotoRepository
