--- app/src/main/java/jumanji/sda/com/jumanji/SavePhotoToDB.kt
@@ -1,30 +1,0 @@
-package jumanji.sda.com.jumanji
-
-import android.net.Uri
-import android.support.annotation.NonNull
-import com.google.android.gms.tasks.OnFailureListener
-import com.google.firebase.storage.UploadTask
-import com.google.android.gms.tasks.OnSuccessListener
-import com.google.firebase.storage.StorageReference
-import java.io.File
-
-
-class SavePhotoToDB {
-
-/*    fun savePhoto() {
-
-        val file = Uri.fromFile(File("path/to/images/rivers.jpg"))
-        val riversRef = storageRef.child("images/rivers.jpg")
-
-        riversRef.putFile(file)
-                .addOnSuccessListener(OnSuccessListener<UploadTask.TaskSnapshot> { taskSnapshot ->
-                    // Get a URL to the uploaded content
-                    val downloadUrl = taskSnapshot.downloadUrl
-                })
-                .addOnFailureListener(OnFailureListener {
-                    // Handle unsuccessful uploads
-                    // ...
-                })
-    }*/
-
-}
